split component
